requirements.txt
fastapi
uvicorn
python-dotenv
pandas
chromadb
ipython
langgraph
langchain
langchain-google-genai
langchain-openai
langchain-community
langchain-ollama
langchain_chroma
langgraph-checkpoint-redis
